Adds in 2016 day 7 part 1

+++ 2015-25_test.go
@@ -6,7 +6,7 @@ func Test2015_25_1(t *testing.T) {
 	cases := []struct {
 		row int
 		col int
-		val int
+		val int64
 	}{
 		{1, 1, 20151125},
 		{4, 3, 21345942},
